+++ include/aquarius/client.hpp
@@ -1,14 +1,12 @@
 ﻿#pragma once
+#include <aquarius/core/asio.hpp>
+#include <aquarius/core/concepts.hpp>
 #include <aquarius/invoke/invoke_callback.hpp>
 #include <aquarius/logger.hpp>
 #include <aquarius/response.hpp>
-#include <aquarius/system/asio.hpp>
-#include <aquarius/system/type_traits.hpp>
 #include <filesystem>
 #include <iostream>
 #include <map>
-#include <type_traits>
-
 
 namespace aquarius
 {
