+++ include/aquarius/core/empty_value.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include <boost/core/empty_value.hpp>
+#include <cstddef>
+
+namespace aquarius
+{
+	template <typename _Alloc>
+	using empty_value = boost::empty_::empty_value<_Alloc>;
+
+	inline static constexpr boost::empty_init_t empty_init = boost::empty_init;
+} // namespace aquarius
