--- include/aquarius/router/impl/singleton.hpp
@@ -1,39 +1,0 @@
-﻿#pragma once
-
-namespace aquarius
-{
-	namespace impl
-	{
-		template <typename _Ty>
-		class singleton_wrapper : public _Ty
-		{};
-
-		template <typename _Ty>
-		class singleton
-		{
-		public:
-			static _Ty& instance()
-			{
-				return get_instance();
-			}
-
-		private:
-			static _Ty& get_instance()
-			{
-				static singleton_wrapper<_Ty> t;
-
-				return static_cast<_Ty&>(t);
-			}
-
-		protected:
-			singleton()
-			{}
-
-		private:
-			static _Ty* instance_;
-		};
-	} // namespace system
-} // namespace aquarius
-
-template <typename _Ty>
-_Ty* aquarius::impl::singleton<_Ty>::instance_ = &singleton<_Ty>::instance();
