+++ include/aquarius/server.hpp
@@ -1,8 +1,7 @@
 #pragma once
 #include <aquarius/invoke/invoke_service.hpp>
 #include <aquarius/invoke/invoke_session.hpp>
-#include <aquarius/system/deadline_timer.hpp>
-#include <type_traits>
+#include <aquarius/core/deadline_timer.hpp>
 
 namespace aquarius
 {
@@ -161,6 +160,6 @@ namespace aquarius
 
 		std::string server_name_;
 
-		system::deadline_timer timer_;
+		deadline_timer timer_;
 	};
 } // namespace aquarius
