+++ include/aquarius/system/visitor.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <aquarius/elastic.hpp>
-#include <aquarius/error_code.hpp>
+#include <aquarius/core/error_code.hpp>
 
 namespace aquarius
 {
