+++ test/unit_test.vcxproj
@@ -158,6 +158,12 @@
     <ClInclude Include="..\include\aquarius\connect.hpp" />
     <ClInclude Include="..\include\aquarius\context.hpp" />
     <ClInclude Include="..\include\aquarius\context\basic_context.hpp" />
+    <ClInclude Include="..\include\aquarius\core\error_code.hpp" />
+    <ClInclude Include="..\include\aquarius\core\impl\error_category.hpp" />
+    <ClInclude Include="..\include\aquarius\core\impl\system_errc.hpp" />
+    <ClInclude Include="..\include\aquarius\core\impl\system_error_message.hpp" />
+    <ClInclude Include="..\include\aquarius\core\impl\time_traits.hpp" />
+    <ClInclude Include="..\include\aquarius\core\impl\type_traits.hpp" />
     <ClInclude Include="..\include\aquarius\invoke\invoke_callback.hpp" />
     <ClInclude Include="..\include\aquarius\message\header.hpp" />
     <ClInclude Include="..\include\aquarius\message\impl\custom.hpp" />
@@ -171,14 +177,13 @@
     <ClInclude Include="..\include\aquarius\router\impl\generator.hpp" />
     <ClInclude Include="..\include\aquarius\router\impl\manager.hpp" />
     <ClInclude Include="..\include\aquarius\router\impl\router.hpp" />
-    <ClInclude Include="..\include\aquarius\router\impl\singleton.hpp" />
+    <ClInclude Include="..\include\aquarius\core\singleton.hpp" />
     <ClInclude Include="..\include\aquarius\router\message.hpp" />
     <ClInclude Include="..\include\aquarius\router\service.hpp" />
     <ClInclude Include="..\include\aquarius\router\session.hpp" />
     <ClInclude Include="..\include\aquarius\session\xsession.hpp" />
-    <ClInclude Include="..\include\aquarius\system\defines.hpp" />
+    <ClInclude Include="..\include\aquarius\core\core.hpp" />
     <ClInclude Include="..\include\aquarius\elastic.hpp" />
-    <ClInclude Include="..\include\aquarius\error_code.hpp" />
     <ClInclude Include="..\include\aquarius\invoke\invoke_context.hpp" />
     <ClInclude Include="..\include\aquarius\invoke\invoke_message.hpp" />
     <ClInclude Include="..\include\aquarius\invoke\invoke_service.hpp" />
@@ -191,17 +196,14 @@
     <ClInclude Include="..\include\aquarius\server.hpp" />
     <ClInclude Include="..\include\aquarius\service.hpp" />
     <ClInclude Include="..\include\aquarius\session.hpp" />
-    <ClInclude Include="..\include\aquarius\system\asio.hpp" />
-    <ClInclude Include="..\include\aquarius\system\deadline_timer.hpp" />
-    <ClInclude Include="..\include\aquarius\system\empty_value.hpp" />
-    <ClInclude Include="..\include\aquarius\system\error_code\error_category.hpp" />
-    <ClInclude Include="..\include\aquarius\system\error_code\system_errc.hpp" />
-    <ClInclude Include="..\include\aquarius\system\error_code\system_error_message.hpp" />
+    <ClInclude Include="..\include\aquarius\core\asio.hpp" />
+    <ClInclude Include="..\include\aquarius\core\deadline_timer.hpp" />
+    <ClInclude Include="..\include\aquarius\core\empty_value.hpp" />
     <ClInclude Include="..\include\aquarius\system\event.hpp" />
     <ClInclude Include="..\include\aquarius\message\field.hpp" />
-    <ClInclude Include="..\include\aquarius\system\uuid.hpp" />
+    <ClInclude Include="..\include\aquarius\core\uuid.hpp" />
     <ClInclude Include="..\include\aquarius\system\visitor.hpp" />
-    <ClInclude Include="..\include\aquarius\system\type_traits.hpp" />
+    <ClInclude Include="..\include\aquarius\core\concepts.hpp" />
     <ClInclude Include="context.h" />
     <ClInclude Include="error_code.h" />
     <ClInclude Include="invoke.h" />
