+++ test/unit_test.vcxproj.filters
@@ -28,9 +28,6 @@
     <Filter Include="include\aquarius\system">
       <UniqueIdentifier>{77e47d6a-d4e0-4016-b765-e7e5d8f33e56}</UniqueIdentifier>
     </Filter>
-    <Filter Include="include\aquarius\system\error_code">
-      <UniqueIdentifier>{cbb83a22-ae46-4acc-973e-de9a23b6f96d}</UniqueIdentifier>
-    </Filter>
     <Filter Include="include\aquarius\session">
       <UniqueIdentifier>{7db6ff92-099d-452e-be84-d69c36512274}</UniqueIdentifier>
     </Filter>
@@ -45,6 +42,12 @@
     </Filter>
     <Filter Include="include\aquarius\router\impl">
       <UniqueIdentifier>{e73a9760-236e-4b21-8fc9-2fa8a7fc5876}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="include\aquarius\core">
+      <UniqueIdentifier>{cc59cd76-2135-40a8-83d7-0ae05a9b6d2b}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="include\aquarius\core\impl">
+      <UniqueIdentifier>{8cb9a4dc-739c-433e-9790-15f7f383bb74}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -122,29 +125,8 @@
     <ClInclude Include="..\include\aquarius\session.hpp">
       <Filter>include\aquarius</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\aquarius\error_code.hpp">
-      <Filter>include\aquarius</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\error_code\error_category.hpp">
-      <Filter>include\aquarius\system\error_code</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\error_code\system_error_message.hpp">
-      <Filter>include\aquarius\system\error_code</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\error_code\system_errc.hpp">
-      <Filter>include\aquarius\system\error_code</Filter>
-    </ClInclude>
     <ClInclude Include="error_code.h">
       <Filter>头文件</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\asio.hpp">
-      <Filter>include\aquarius\system</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\deadline_timer.hpp">
-      <Filter>include\aquarius\system</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\empty_value.hpp">
-      <Filter>include\aquarius\system</Filter>
     </ClInclude>
     <ClInclude Include="..\include\aquarius\system\event.hpp">
       <Filter>include\aquarius\system</Filter>
@@ -155,17 +137,8 @@
     <ClInclude Include="..\include\aquarius\service.hpp">
       <Filter>include\aquarius</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\defines.hpp">
-      <Filter>include\aquarius\system</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\type_traits.hpp">
-      <Filter>include\aquarius\system</Filter>
-    </ClInclude>
     <ClInclude Include="..\include\aquarius\invoke\invoke_callback.hpp">
       <Filter>include\aquarius\invoke</Filter>
-    </ClInclude>
-    <ClInclude Include="..\include\aquarius\system\uuid.hpp">
-      <Filter>include\aquarius\system</Filter>
     </ClInclude>
     <ClInclude Include="..\include\aquarius\session\xsession.hpp">
       <Filter>include\aquarius\session</Filter>
@@ -209,9 +182,6 @@
     <ClInclude Include="..\include\aquarius\router\impl\router.hpp">
       <Filter>include\aquarius\router\impl</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\aquarius\router\impl\singleton.hpp">
-      <Filter>include\aquarius\router\impl</Filter>
-    </ClInclude>
     <ClInclude Include="..\include\aquarius\router\impl\generator.hpp">
       <Filter>include\aquarius\router\impl</Filter>
     </ClInclude>
@@ -223,6 +193,45 @@
     </ClInclude>
     <ClInclude Include="..\include\aquarius\router\impl\apply.hpp">
       <Filter>include\aquarius\router\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\asio.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\core.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\deadline_timer.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\empty_value.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\concepts.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\uuid.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\singleton.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\impl\time_traits.hpp">
+      <Filter>include\aquarius\core\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\impl\type_traits.hpp">
+      <Filter>include\aquarius\core\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\error_code.hpp">
+      <Filter>include\aquarius\core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\impl\error_category.hpp">
+      <Filter>include\aquarius\core\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\impl\system_errc.hpp">
+      <Filter>include\aquarius\core\impl</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\aquarius\core\impl\system_error_message.hpp">
+      <Filter>include\aquarius\core\impl</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
